Issue 83 - Upgrade to 2017 WPILib
Upgrade to the 2017 version of WPILib for Java, which is currently only available to those in the Beta program but will be generally available on January 7, 2017. This version removes the `CANTalon` class and instead replaces it with a 3rd party library from Cross The Road Electronics (CTRE), who is now going to be responsible for that class.

The CTRE library is also currently accessible only to those teams in the Beta program, so the build installs it from a local file during the `deps` Ant build target. This is specified in the `build.properties` file and will be changed once we have a public URL for the download.