Included in archives the 3rd party libraries needed for robot
The CTRE and navX-MXP libraries are needed on the robot at runtime, and they are now included in the Strongback TAR and ZIP archive in the `strongback/java/lib` directory (next to the `strongback.jar` and `strongback-sources.jar` files).

The Strongback CLI will need to be modified to copy these into the WPILib’s `user/lib/java` directory, though it should optionally not overwrite the non-Strongback files.